chore: rename git module to git_api
GitPython's toplevel module is also called `git`. This caused import issues in some
situations (e.g. when running `__main__.py`)
