+++ gitopscli/commands/add_pr_comment.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from typing import Optional
-from gitopscli.git import GitApiConfig, GitRepoApiFactory
+from gitopscli.git_api import GitApiConfig, GitRepoApiFactory
 from .command import Command
