+++ gitopscli/commands/deploy.py
@@ -2,7 +2,7 @@ import logging
 import uuid
 from dataclasses import dataclass
 from typing import Any, Dict, Optional, Tuple
-from gitopscli.git import GitApiConfig, GitRepo, GitRepoApi, GitRepoApiFactory
+from gitopscli.git_api import GitApiConfig, GitRepo, GitRepoApi, GitRepoApiFactory
 from gitopscli.io.yaml_util import update_yaml_file, yaml_dump, YAMLException
 from gitopscli.gitops_exception import GitOpsException
 from .command import Command
