+++ gitopscli/git_api/bitbucket_git_repo_api_adapter.py
