+++ tests/commands/common/test_gitops_config_loader.py
@@ -4,7 +4,7 @@ import pytest
 from gitopscli.gitops_exception import GitOpsException
 from gitopscli.gitops_config import GitOpsConfig
 from gitopscli.io.yaml_util import yaml_file_load
-from gitopscli.git import GitApiConfig, GitProvider, GitRepo, GitRepoApi, GitRepoApiFactory
+from gitopscli.git_api import GitApiConfig, GitProvider, GitRepo, GitRepoApi, GitRepoApiFactory
 from gitopscli.commands.common.gitops_config_loader import load_gitops_config
 from tests.commands.mock_mixin import MockMixin
