+++ tests/git_api/test_git_repo_api_logging_proxy.py
@@ -1,8 +1,8 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from gitopscli.git import GitRepoApi
-from gitopscli.git.git_repo_api_logging_proxy import GitRepoApiLoggingProxy
+from gitopscli.git_api import GitRepoApi
+from gitopscli.git_api.git_repo_api_logging_proxy import GitRepoApiLoggingProxy
 
 
 class GitRepoApiLoggingProxyTest(unittest.TestCase):
@@ -37,7 +37,7 @@ class GitRepoApiLoggingProxyTest(unittest.TestCase):
         self.assertEqual(actual_return_value, expected_return_value)
         self.__mock_repo_api.get_clone_url.assert_called_once_with()
 
-    @patch("gitopscli.git.git_repo_api_logging_proxy.logging")
+    @patch("gitopscli.git_api.git_repo_api_logging_proxy.logging")
     def test_create_pull_request(self, logging_mock):
         expected_return_value = GitRepoApi.PullRequestIdAndUrl(42, "<url>")
         self.__mock_repo_api.create_pull_request.return_value = expected_return_value
@@ -54,7 +54,7 @@ class GitRepoApiLoggingProxyTest(unittest.TestCase):
             "Creating pull request from '%s' to '%s' with title: %s", "<from branch>", "<to branch>", "<title>",
         )
 
-    @patch("gitopscli.git.git_repo_api_logging_proxy.logging")
+    @patch("gitopscli.git_api.git_repo_api_logging_proxy.logging")
     def test_create_pull_request_to_default_branch(self, logging_mock):
         expected_return_value = GitRepoApi.PullRequestIdAndUrl(42, "<url>")
         self.__mock_repo_api.create_pull_request_to_default_branch.return_value = expected_return_value
@@ -71,7 +71,7 @@ class GitRepoApiLoggingProxyTest(unittest.TestCase):
             "Creating pull request from '%s' to default branch with title: %s", "<from branch>", "<title>",
         )
 
-    @patch("gitopscli.git.git_repo_api_logging_proxy.logging")
+    @patch("gitopscli.git_api.git_repo_api_logging_proxy.logging")
     def test_merge_pull_request(self, logging_mock):
         self.__testee.merge_pull_request(pr_id=42)
         self.__mock_repo_api.merge_pull_request.assert_called_once_with(42)
@@ -79,7 +79,7 @@ class GitRepoApiLoggingProxyTest(unittest.TestCase):
             "Merging pull request %s", 42,
         )
 
-    @patch("gitopscli.git.git_repo_api_logging_proxy.logging")
+    @patch("gitopscli.git_api.git_repo_api_logging_proxy.logging")
     def test_add_pull_request_comment(self, logging_mock):
         self.__testee.add_pull_request_comment(pr_id=42, text="<text>", parent_id=4711)
         self.__mock_repo_api.add_pull_request_comment.assert_called_once_with(42, "<text>", 4711)
@@ -87,7 +87,7 @@ class GitRepoApiLoggingProxyTest(unittest.TestCase):
             "Creating comment for pull request %s as reply to comment %s with content: %s", 42, 4711, "<text>",
         )
 
-    @patch("gitopscli.git.git_repo_api_logging_proxy.logging")
+    @patch("gitopscli.git_api.git_repo_api_logging_proxy.logging")
     def test_add_pull_request_comment_without_parent_id(self, logging_mock):
         self.__testee.add_pull_request_comment(pr_id=42, text="<text>", parent_id=None)
         self.__mock_repo_api.add_pull_request_comment.assert_called_once_with(42, "<text>", None)
@@ -95,7 +95,7 @@ class GitRepoApiLoggingProxyTest(unittest.TestCase):
             "Creating comment for pull request %s with content: %s", 42, "<text>",
         )
 
-    @patch("gitopscli.git.git_repo_api_logging_proxy.logging")
+    @patch("gitopscli.git_api.git_repo_api_logging_proxy.logging")
     def test_delete_branch(self, logging_mock):
         self.__testee.delete_branch("<branch>")
         self.__mock_repo_api.delete_branch.assert_called_once_with("<branch>")
